[#12] feature : fix style
