+++ FE/storybook/stories/JyeongMu/src/components/loginHeader.jsx
@@ -1,27 +1,9 @@
 /* eslint-disable react/destructuring-assignment */
 /* eslint-disable react/prop-types */
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import { Text, View } from 'react-native';
+import styles from '../style/style';
 
-const styles = StyleSheet.create({
-  containter: {
-    flex: 1,
-    padding: '5%',
-    alignItems: 'center',
-  },
-  header: {
-    flex: 1,
-    alignItems: 'center',
-    height: '7%',
-    paddingTop: '20%',
-  },
-  headerText: {
-    fontSize: 30,
-    fontStyle: 'italic',
-    color: 'black',
-  },
-
-});
 export default function LoginHeader(props) {
   return (
     <View style={styles.header}>
