+++ FE/storybook/stories/JyeongMu/src/style/style.jsx
@@ -1,0 +1,56 @@
+import { StyleSheet } from 'react-native';
+
+// eslint-disable-next-line no-unused-vars
+const styles = StyleSheet.create({
+  containter: {
+    flex: 1,
+    alignItems: 'center',
+    flexDirection: 'row',
+  },
+  selectButton: {
+    flex: 1,
+    alignItems: 'center',
+  },
+  header: {
+    flex: 1,
+    alignItems: 'center',
+    height: '7%',
+    padding: '5%',
+    margin: '10%',
+  },
+  headerText: {
+    fontSize: 30,
+    fontStyle: 'italic',
+    color: 'black',
+  },
+  text: {
+    fontSize: 15,
+  },
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    flexDirection: 'column',
+  },
+  body: {
+    flex: 9,
+    alignItems: 'center',
+  },
+  select: {
+    flexDirection: 'row',
+    margin: '5%',
+    justifyContent: 'space-around',
+  },
+  normalBody: {
+    margin: '3%',
+  },
+  textInput: {
+    padding: '2%',
+    margin: '5%',
+    borderRadius: 10,
+    backgroundColor: '#eeeeee',
+    width: '80%',
+  },
+  buttonContainer: {
+    width: '80%',
+  },
+});
