+++ FE/storybook/stories/index.jsx
@@ -12,7 +12,6 @@ import Welcome from './Welcome';
 import Something from './components/Something';
 
 import Login from './JyeongMu/src/screen/login';
-import PracticeStyleComponent from './JyeongMu/src/components/practice_style_component';
 
 storiesOf('Welcome', module).add('to Storybook', () => <Welcome showApp={linkTo('Button')} />);
 
@@ -33,5 +32,4 @@ storiesOf('Button', module)
   ));
 
 storiesOf('정무', module)
-  .add('로그인 페이지', () => <Login showApp={linkTo('Button')} />)
-  .add('스타일 컴포넌트 연습', () => <PracticeStyleComponent showApp={linkTo('Button')} />);
+  .add('로그인 페이지', () => <Login showApp={linkTo('Button')} />);
